requirements.txt
aiobotocore==2.11.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
botocore==1.34.27
frozenlist==1.4.1
fsspec==2023.12.2
idna==3.6
jmespath==1.0.1
kafka-python==2.0.2
multidict==6.0.4
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
s3fs==2023.12.2
six==1.16.0
tzdata==2023.4
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.4
